QR and Barcode implementation issue resolved
